lx2162aqds: re-enable dpmac11
In the LX2162A DEVDISR2, bits 9:6 are reserved, but it's not hard to
imagine what they are, since the surrounding bits 5:0 are MAC6..MAC1 and
bits 17:10 are MAC18..MAC11.

It seems the hidden MAC10..MAC7 have to be turned off in the PBL, on
account of the fact that these MACs are missing/unconnected in LX2162A.
The SerDes1 of this SoC has fewer lanes than LX2160A (and on LX2160A,
that's where these MACs were routed to).

Disabling unused MACs 7-10 is fine, but the problem is that the written
bit field of 0x7c0 is too wide for the intended purpose and also
disables MAC11 - an SGMII MAC routinely connected to SerDes2 lane A.
This was most certainly unintended, given that the disable_mac7_10.rcw
PBI fragment applies by default in all LX2162A RCW files.

Fix the initializer value of DEVDISR2 to only disable dpmac7-10.

Fixes: 7911cbebbe91 ("LX2: Add support of LX2162AQDS board")
Signed-off-by: Josua Mayer <josua@solid-run.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>