rcw: add loadc, jumpc and jump to pbi instructions
Add processing for the 'load conditional', 'jump conditional' and 'jump'
instructions.

These are all Gen3 architecture instructions (parsed by the Service
Processor as in LX2160A), so we always need to test for %pbiformat=2.

Commit 7c47f3027817 ("Add support of Gen3 family SoCs") has left some
stubs for disassembling these instructions without ever leaving code for
assembling them.

As there have already appeared users in the wild of the 'load
conditional' instruction (notably the Solidrun auto-boot feature,
which includes support for multiple boot sources in the same RCW binary,
and selects among them based on the PORSR1[RCW_SRC] field) - it makes
sense not to break what those users have already set up, and that is to
name the instruction 'loadc' (since that naming is what they chose).

[ vjardin: update 'loadcondition' disassembly to 'loadc', to match
  human-generated code ]
[ vladimiroltean: implement missing jump and jumpc disassembly;
  refactor to dedicated functions per instruction ]

Signed-off-by: Rabeeh Khoury <rabeeh@solid-run.com>
Co-developed-by: Vincent Jardin <vjardin@free.fr>
Signed-off-by: Vincent Jardin <vjardin@free.fr>
Co-developed-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>